feat(overview): chart support time range selection
